Renamed CurrentMetrics to Metrics and placed metrics snapshot acquire method at ZIO object.
